CAMEL-21699: camel-kafka - Batch consumer should not require message body to keep being List<Exchange>
