+++ components/camel-kafka/src/main/java/org/apache/camel/component/kafka/KafkaProducer.java
@@ -41,18 +41,15 @@ import org.apache.camel.health.HealthCheckHelper;
 import org.apache.camel.health.WritableHealthCheckRepository;
 import org.apache.camel.spi.HeaderFilterStrategy;
 import org.apache.camel.support.DefaultAsyncProducer;
-import org.apache.camel.support.SynchronizationAdapter;
 import org.apache.camel.util.KeyValueHolder;
 import org.apache.camel.util.ObjectHelper;
 import org.apache.camel.util.ReflectionHelper;
 import org.apache.camel.util.URISupport;
 import org.apache.kafka.clients.CommonClientConfigs;
 import org.apache.kafka.clients.NetworkClient;
-import org.apache.kafka.clients.producer.Producer;
 import org.apache.kafka.clients.producer.ProducerConfig;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.clients.producer.RecordMetadata;
-import org.apache.kafka.common.KafkaException;
 import org.apache.kafka.common.header.Header;
 import org.apache.kafka.common.header.internals.RecordHeader;
 import org.slf4j.Logger;
@@ -527,44 +524,5 @@ public class KafkaProducer extends DefaultAsyncProducer {
         exchange.getUnitOfWork().beginTransactedBy(transactionId);
         kafkaProducer.beginTransaction();
         exchange.getUnitOfWork().addSynchronization(new KafkaTransactionSynchronization(transactionId, kafkaProducer));
-    }
-}
-
-class KafkaTransactionSynchronization extends SynchronizationAdapter {
-    private static final Logger LOG = LoggerFactory.getLogger(KafkaTransactionSynchronization.class);
-    private final String transactionId;
-    private final Producer kafkaProducer;
-
-    public KafkaTransactionSynchronization(String transactionId, Producer kafkaProducer) {
-        this.transactionId = transactionId;
-        this.kafkaProducer = kafkaProducer;
-    }
-
-    @Override
-    public void onDone(Exchange exchange) {
-        try {
-            if (exchange.getException() != null || exchange.isRollbackOnly()) {
-                if (exchange.getException() instanceof KafkaException) {
-                    LOG.warn("Catch {} and will close kafka producer with transaction {} ", exchange.getException(),
-                            transactionId);
-                    kafkaProducer.close();
-                } else {
-                    LOG.warn("Abort kafka transaction {} with exchange {}", transactionId, exchange.getExchangeId());
-                    kafkaProducer.abortTransaction();
-                }
-            } else {
-                LOG.debug("Commit kafka transaction {} with exchange {}", transactionId, exchange.getExchangeId());
-                kafkaProducer.commitTransaction();
-            }
-        } catch (KafkaException e) {
-            exchange.setException(e);
-        } catch (Exception e) {
-            exchange.setException(e);
-            LOG.warn("Abort kafka transaction {} with exchange {} due to {} ", transactionId, exchange.getExchangeId(),
-                    e.getMessage(), e);
-            kafkaProducer.abortTransaction();
-        } finally {
-            exchange.getUnitOfWork().endTransactedBy(transactionId);
-        }
     }
 }
